Rename file; add updates and ed component
Mostly placeholders here.
